Port all DenseVector code to Vector3

+++ DemoMacroFeature/SampleMacroFeature/SampleMacroFeature.cs
@@ -50,7 +50,7 @@ namespace DemoMacroFeatures.SampleMacroFeature
             var axisX = Vector3.UnitX;
 
             // Find the point to cut the object
-            center.X = database.Alpha*box.P0.X + (1 - database.Alpha)*box.P1.X;
+            center.X = (float) (database.Alpha*box.P0.X + (1 - database.Alpha)*box.P1.X);
             var sheet = modeler.CreateSheet(center, axisX, box.P0, box.P1);
 
             var cutResult = body.Cut(sheet);
